qdrant_migrations/requirements.txt
psycopg2-binary
# sentence-transformers
qdrant-client
requests
